Modified NativeKeyCode's 'HOME' keycode value, Changed remoteroid.idc, to make 'remoteroid' device as not 'built-in' type
When 'remoteroid' connected to device, it uses Generic.kcm for key character map and Generic.kl for keylayout.
(I defined kcm and keylayout to remoteroid for each and copied each file into proper place. But system still uses Generic rather than remoteroid, which is my version of kcm and keylayout.)
My version of keylayout is just the same copy of 'qwerty.kl', which differs in 'HOME' keycode in Generic.kl. So I modified keycodes in NativeKeycode.java to fit with Generic.kl. (just changed keycode for 'HOME', see source code for details)